Create two scripts to control players
Two different scripts used instead of one because of the possibility of including AI opponent for single player functionality.
This way, single player code can be added into Player1Controller, so that it is only applicable to player 1 and player 2 is ignored. As well, this makes input managing a little simpler to implement.